add crons options on fort
