+++ src/utils/cron_manager.ts
@@ -1,4 +1,6 @@
 
 export class CronJobManager {
-
+    start(name: string) {
+        
+    }
 }
